all endpoint success done

+++ hayumabar/app/Controllers/Http/VenuesController.ts
@@ -4,9 +4,9 @@ import CreateVenueValidator from "App/Validators/CreateVenueValidator";
 
 export default class VenuesController {
   public async index({ response }: HttpContextContract) {
-    let venues = await Venue.all();
+    let venues = await Venue.query().select("*").preload("fields");
     response.ok({
-      message: "Berhasil get all data venues",
+      message: "Berhasil mendapatkan semua data venues",
       data: venues,
     });
   }
@@ -15,15 +15,16 @@ export default class VenuesController {
 
   public async store({ auth, request, response }: HttpContextContract) {
     const payload = await request.validate(CreateVenueValidator);
-    const authUser = auth.user;
+    const user = auth.user!;
+
     const newVenue = await Venue.create({
       name: payload.name,
       address: payload.address,
       phone: payload.phone,
-      user_id: authUser?.id,
+      userId: user.id,
     });
     response.created({
-      message: "Venue berhasil dibuat",
+      message: "Venue telah berhasil dibuat",
       venueName: newVenue.name,
     });
   }
@@ -31,9 +32,12 @@ export default class VenuesController {
   public async show({ response, params }: HttpContextContract) {
     let venue = await Venue.query()
       .select("*")
-      .preload("fields")
+      .preload("fields", (fieldQuery) => {
+        fieldQuery.preload("bookings");
+      })
       .where("id", params.id)
       .firstOrFail();
+
     response.ok({
       message: "Berhasil mendapatkan data venue dan fields venue",
       data: venue,
@@ -42,23 +46,47 @@ export default class VenuesController {
 
   public async edit({}: HttpContextContract) {}
 
-  public async update({ request, response, params }: HttpContextContract) {
-    const payload = await request.validate(CreateVenueValidator);
-    let venue = await Venue.findOrFail(params.id);
-    venue.name = payload.name;
-    venue.address = payload.address;
-    venue.phone = payload.phone;
-    await venue.save();
+  public async update({
+    auth,
+    request,
+    response,
+    params,
+  }: HttpContextContract) {
+    const user = auth.user!;
+    const cekVenue = await Venue.findByOrFail("id", params.id);
 
-    response.ok({
-      message: "Venue berhasil diperbaharui",
-      updatedVenueName: venue.name,
-    });
+    if (user.id == cekVenue.userId) {
+      const payload = await request.validate(CreateVenueValidator);
+
+      let venue = await Venue.query()
+        .where("user_id", user.id)
+        .andWhere("id", params.venue_id)
+        .select("*")
+        .firstOrFail();
+
+      venue.name = payload.name;
+      venue.address = payload.address;
+      venue.phone = payload.phone;
+
+      await venue.save();
+
+      return response.ok({
+        message: `Venue dengan ID: ${venue.id} telah berhasil diperbaharui`,
+      });
+    } else {
+      return response.unauthorized({
+        message: "Kamu tidak dapat mengakses ini",
+      });
+    }
   }
 
-  public async destroy({ params, response }: HttpContextContract) {
-    let venue = await Venue.findOrFail(params.id);
-    await venue.delete();
-    response.ok({ message: "Venue berhasil dihapus" });
+  public async destroy({ auth, params, response }: HttpContextContract) {
+    const user = auth.user!;
+    const venue = await Venue.findOrFail(params.id);
+
+    if (user.id == venue.userId) {
+      await venue.delete();
+      response.ok({ message: "Venue berhasil dihapus" });
+    }
   }
 }
